refactor(app): wrap dispatch and pass the wrapped dispatch to form
only 'echo' if we use the command word
