+++ src/components/App/App.tsx
@@ -1,76 +1,46 @@
-import React, { FC, Reducer, useReducer } from 'react';
-import { v4 as uuid } from 'uuid';
+import React, { FC, useReducer, useCallback } from 'react';
 
+import { initialState, reducer } from '../../app';
+
+import { Fireworks } from '../Fireworks';
 import { Form } from '../Form';
 import { Messages } from '../Messages';
-import { AppAction, AppState } from '../types';
 
 import style from './App.module.css';
-
-const reducer: Reducer<AppState, AppAction> = (state, action) => {
-  switch (action.type) {
-    case 'setMessagesTop':
-      return {
-        messages: state.messages,
-        messagesTop: action.payload.messagesTop,
-        messagesHeight: state.messagesHeight,
-      };
-    case 'setMessagesHeight':
-      return {
-        messages: state.messages,
-        messagesTop: state.messagesTop,
-        messagesHeight: action.payload.messagesHeight,
-      };
-    case 'addMessage':
-      return {
-        messages: [
-          ...state.messages,
-          {
-            id: uuid(),
-            createdAt: new Date().toISOString(),
-            ...action.payload,
-          },
-        ],
-        messagesTop: state.messagesTop,
-        messagesHeight: state.messagesHeight,
-      };
-    case 'renderMessage':
-      const message = state.messages.find((m) => m.id === action.payload.id);
-      const index = message ? state.messages.indexOf(message) : -1;
-
-      return message && index !== -1
-        ? {
-            messages: [
-              ...state.messages.slice(0, index),
-              {
-                ...message,
-                top: action.payload.top,
-                height: action.payload.height,
-              },
-              ...state.messages.slice(index + 1),
-            ],
-            messagesTop: state.messagesTop,
-            messagesHeight: state.messagesHeight + action.payload.height,
-          }
-        : state;
-    default:
-      return state;
-  }
-};
+import { AppAction, Sender } from '../../types';
 
 export const App: FC = () => {
   const [{ messages, messagesTop, messagesHeight }, dispatch] = useReducer(
     reducer,
-    {
-      messages: [],
-      messagesTop: 0,
-      messagesHeight: 0,
+    initialState,
+  );
+
+  const wrappedDispatch = useCallback(
+    (action: AppAction) => {
+      dispatch(action);
+
+      if (
+        action.type === 'addMessage' &&
+        action.payload.content.indexOf('say ') === 0
+      ) {
+        setTimeout(() => {
+          dispatch({
+            type: 'addMessage',
+            payload: {
+              sender: Sender.Other,
+              content: action.payload.content.slice(3).trim(),
+            },
+          });
+        }, 2_000);
+      }
     },
+    [dispatch],
   );
 
   return (
     <article className={style.Device}>
-      <header className={style.Header}>Echo</header>
+      <Fireworks />
+      <header className={style.Header}>Bot</header>
       <Messages
         messages={messages}
         messagesTop={messagesTop}
@@ -78,7 +48,7 @@ export const App: FC = () => {
         dispatch={dispatch}
       />
       <footer className={style.Footer}>
-        <Form dispatch={dispatch} />
+        <Form dispatch={wrappedDispatch} />
       </footer>
     </article>
   );
